Finaliza alterações do modelo
Foi necessário mover o model edital para a app contratos
uma vez que agora faz parte do mesmo assunto e para
evitar problemas de import circilar.
