+++ sme_coad_apps/contratos/admin.py
@@ -8,11 +8,14 @@ from .models import (
     Contrato,
     ContratoUnidade,
     DotacaoValor,
+    Edital,
     Empresa,
     FiscaisUnidade,
     FiscalLote,
+    GrupoObrigacao,
     Lote,
     NotificacaoVigenciaContrato,
+    Obrigacao,
     ObrigacaoContratual,
     ParametroNotificacoesVigencia,
     TipoServico
@@ -211,3 +214,38 @@ class LoteAdmin(admin.ModelAdmin):
 class DotacaoValorAdmin(admin.ModelAdmin):
     list_display = ['contrato', 'dotacao_orcamentaria', 'valor', ]
     ordering = ('contrato',)
+
+
+class GrupoObrigacoesInline(admin.StackedInline):
+    extra = 1
+    model = GrupoObrigacao
+
+
+class ObrigacaoInline(admin.StackedInline):
+    extra = 1
+    model = Obrigacao
+    fieldsets = (
+        (None, {
+            'fields': ('item', 'descricao')
+        }),
+    )
+
+
+@admin.register(Edital)
+class EditalAdmin(admin.ModelAdmin):
+    list_display = ('numero',)
+    ordering = ('numero',)
+    search_fields = ('numero',)
+    inlines = [
+        GrupoObrigacoesInline
+    ]
+
+
+@admin.register(GrupoObrigacao)
+class GrupoObrigacaoAdmin(admin.ModelAdmin):
+    list_display = ('nome',)
+    ordering = ('nome',)
+    search_fields = ('nome',)
+    inlines = [
+        ObrigacaoInline
+    ]
