+++ sme_coad_apps/contratos/api/serializers/grupo_obrigacao_serializer.py
