+++ sme_coad_apps/contratos/models/__init__.py
@@ -1,8 +1,10 @@
 from .colunas_contrato import ColunasContrato
 from .contrato import Contrato, ContratoUnidade, FiscaisUnidade, FiscalLote, Lote
 from .dotacao_valor import DotacaoValor
+from .edital import Edital
 from .empresa import Empresa
 from .notificacao_vigencia_contrato import NotificacaoVigenciaContrato
+from .obrigacao import GrupoObrigacao, Obrigacao
 from .obrigacao_contratual import ObrigacaoContratual
 from .parametro_notificacoes import ParametroNotificacoesVigencia
 from .tipo_servico import TipoServico
