+++ sme_coad_apps/contratos/tests/test_contrato_serializer.py
@@ -4,9 +4,9 @@ import pytest
 from model_mommy import mommy
 
 from ...atestes.models import ModeloAteste
-from ...core.models import Edital, Nucleo
+from ...core.models import Nucleo
 from ..api.serializers.contrato_serializer import ContratoLookUpSerializer, ContratoSerializer
-from ..models.contrato import Contrato, Empresa, TipoServico
+from ..models.contrato import Contrato, Empresa, Edital, TipoServico
 
 pytestmark = pytest.mark.django_db
