+++ sme_coad_apps/core/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import Coad, CoadAssessor, Divisao, Edital, GrupoObrigacao, Nucleo, Obrigacao, Servidor, Unidade
+from .models import Coad, CoadAssessor, Divisao, Nucleo, Servidor, Unidade
 
 
 class NucleoInLine(admin.TabularInline):
@@ -53,38 +53,3 @@ class CoadAdmin(admin.ModelAdmin):
 @admin.register(Servidor)
 class ServidorAdmin(admin.ModelAdmin):
     ...
-
-
-class GrupoObrigacoesInline(admin.StackedInline):
-    extra = 1
-    model = GrupoObrigacao
-
-
-class ObrigacaoInline(admin.StackedInline):
-    extra = 1
-    model = Obrigacao
-    fieldsets = (
-        (None, {
-            'fields': ('item', 'descricao')
-        }),
-    )
-
-
-@admin.register(Edital)
-class EditalAdmin(admin.ModelAdmin):
-    list_display = ('numero',)
-    ordering = ('numero',)
-    search_fields = ('numero',)
-    inlines = [
-        GrupoObrigacoesInline
-    ]
-
-
-@admin.register(GrupoObrigacao)
-class GrupoObrigacaoAdmin(admin.ModelAdmin):
-    list_display = ('nome',)
-    ordering = ('nome',)
-    search_fields = ('nome',)
-    inlines = [
-        ObrigacaoInline
-    ]
