+++ sme_coad_apps/core/models/__init__.py
@@ -1,8 +1,6 @@
 from .coad import Coad
 from .coad_assessor import CoadAssessor
 from .divisao import Divisao
-from .edital import Edital
 from .nucleo import Nucleo
-from .obrigacao import GrupoObrigacao, Obrigacao
 from .servidor import Servidor
 from .unidade import Unidade
