+++ sme_coad_apps/core/urls.py
@@ -6,7 +6,6 @@ from sme_coad_apps.contratos.urls import router as contratos_router
 from sme_coad_apps.users.urls import router as usuarios_router
 
 from ..contratos.models.contrato import SafiToken
-from ..core.api.viewsets.edital_viewset import EditalViewSet
 from ..core.api.viewsets.unidade_viewset import UnidadeViewSet
 from .api.viewsets.coad_assessor_viewset import CoadAssessorViewSet
 from .api.viewsets.coad_viewset import CoadViewSet
@@ -20,7 +19,6 @@ router.register('nucleos', NucleoViewSet)
 router.register('coad', CoadViewSet)
 router.register('coad-assessores', CoadAssessorViewSet)
 router.register('unidades', UnidadeViewSet)
-router.register('editais', EditalViewSet)
 router.register('safi-token', SafiToken, basename='Version')
 
 router.registry.extend(contratos_router.registry)
